testing/web-platform/tests/atomics/dedicated-worker-agent/wait/good-views.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/workerhelper.js></script>
<script>

/*---
description: >
  Test Atomics.wait on arrays that allow atomic operations,
  in an Agent that is allowed to wait.
---*/

window.addEventListener("load", function() {
  async_test(function(t) {
    var check_done = WorkerHelper.check_done(t, 7);

    var A = WorkerHelper.createFromString(`
        onmessage = function(e) {
            var view = e.data[0];
            postMessage(Atomics.wait(view, 0, 0, 0));
        }
    `, t.step_func(function(e) {
        assert_equals(e.data, "timed-out");
        console.log("A");
        check_done();
    }));
    var B = WorkerHelper.createFromString(`
        onmessage = function(e) {
            var view = e.data[0];
            postMessage(Atomics.wait(view, 0, 37, 0));
        }
    `, t.step_func(function(e) {
        assert_equals(e.data, "not-equal");
        check_done();
    }));
    var C = WorkerHelper.createFromString(`
        onmessage = function(e) {
            var view = e.data[0];
            var idx = e.data[1];
            postMessage(Atomics.wait(view, idx, 0));
        }
    `, t.step_func(function(e) {
        assert_equals(e.data, "not-equal");
        check_done();
    }));

    var sab = new SharedArrayBuffer(1024);

    var good_indices = [
        (view) => 0/-1,
        (view) => '-0',
        (view) => view.length - 1,
        (view) => ({ valueOf: () => 0 }),
        (view) => ({ toString: () => '0', valueOf: false }) // non-callable valueOf triggers invocation of toString
    ];

    var view = new Int32Array(sab, 32, 20);

    view[0] = 0;
    A.postMessage([view]);
    B.postMessage([view]);

    setTimeout(function() {
        for (var IdxGen of good_indices) {
            let Idx = IdxGen(view);
            view[Idx] = 0;
            // Firefox cannot apply clone algorithm on an Object.
            // See https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm
            if (typeof Idx == "object") {
                Idx = Idx.valueOf ? Idx.valueOf() : Idx.toString();
            }
            Atomics.store(view, Idx, 37);
            C.postMessage([view, Idx]);
        }
    }, 500);
  }, "atomics-dedicated-worker-wait-good-views");
});

</script>
